# openapi-snapshot.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "HLOOL Mail API",
    "version": "0.8.0",
    "description": "由内部 apispec 注册表生成的机器可读 API 元数据。"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "已配置的公开基础 URL"
    }
  ],
...
  "paths": {
    "/api/docs.md": {
      "get": {
        "tags": [
          "公开元数据"
        ],
        "summary": "Markdown API 指南",
        "description": "返回由 API 注册表生成、适合 AI 阅读的 Markdown 指南。",
        "operationId": "apiDocsMarkdown",
        "responses": {
          "200": {
            "description": "Markdown 文档",
            "content": {
              "text/markdown": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "public"
      }
    },
    "/api/domains/available": {
      "get": {
        "tags": [
          "API Key 自动化"
        ],
        "summary": "可用域名",
        "description": "列出可选择的公共域名以及 API Key 拥有者可访问的私有域名。客户端和 AI 助手应把该端点作为生成邮箱前的域名选择入口，优先使用 data.public_domains 与 data.private_domains；data.domains 仅为旧客户端公共域名兼容字段。",
        "operationId": "availableDomains",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeAvailableDomains"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "X-API-Key"
      }
    },
    "/api/email/{id}": {
      "delete": {
        "tags": [
          "API Key 自动化"
        ],
        "summary": "删除邮件",
        "description": "删除 API Key 调用方有权限访问的一封邮件。",
        "operationId": "deleteEmail",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "邮件 ID。",
            "schema": {
              "type": "string"
            },
            "example": "msg-uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeDeleteMessage"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "X-API-Key"
      },
      "get": {
        "tags": [
          "API Key 自动化"
        ],
        "summary": "读取邮件",
        "description": "按 ID 读取一封邮件，包含文本正文、可用时的安全 HTML、邮件头和附件元数据。",
        "operationId": "getEmail",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "邮件 ID。",
            "schema": {
              "type": "string"
            },
            "example": "msg-uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeMessageDetail"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "X-API-Key"
      }
    },
    "/api/email/{id}/read": {
      "patch": {
        "tags": [
          "API Key 自动化"
        ],
        "summary": "标记邮件已读",
        "description": "将一封邮件标记为已读。",
        "operationId": "markEmailRead",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "邮件 ID。",
            "schema": {
              "type": "string"
            },
            "example": "msg-uuid"
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeMarkRead"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "X-API-Key"
      }
    },
    "/api/emails": {
      "get": {
        "tags": [
          "API Key 自动化"
        ],
        "summary": "邮件列表",
        "description": "列出某个邮箱的邮件，不改变已读状态。使用 page 和 per_page 获取分页元数据。",
        "operationId": "listEmails",
        "security": [
          {
            "apiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "email",
            "in": "query",
            "required": true,
            "description": "邮箱地址。",
            "schema": {
              "format": "email",
              "type": "string"
            },
            "example": "verify@example.com"
          },
          {
            "name": "page",
            "in": "query",
            "description": "分页页码。",
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "example": 1
          },
          {
            "name": "per_page",
            "in": "query",
            "description": "每页数量。",
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "example": 20
          },
          {
            "name": "limit",
            "in": "query",
            "description": "兼容旧数组响应的数量限制。",
            "schema": {
              "format": "int32",
              "type": "integer"
            },
            "example": 50
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EnvelopeMessagePage"
                }
              }
            }
          },
          "400": {
            "description": "请求参数错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "401": {
            "description": "未认证",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "403": {
            "description": "无权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "未找到",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "410": {
            "description": "资源已失效",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "429": {
            "description": "请求过于频繁",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "服务器内部错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        },
        "x-hlool-auth": "X-API-Key"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeleteMessageData": {
        "additionalProperties": false,
        "properties": {
          "deleted": {
            "type": "boolean"
          }
        },
        "required": [
          "deleted"
        ],
        "type": "object"
      },
...
      "EnvelopeAvailableDomains": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/AvailableDomainsData"
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
...
      "EnvelopeDeleteMessage": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeleteMessageData"
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
...
      "EnvelopeMarkRead": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MarkReadData"
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
      "EnvelopeMessageDetail": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MessageDetail"
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
      "EnvelopeMessagePage": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "$ref": "#/components/schemas/MessagePage"
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
...
      "ErrorEnvelope": {
        "additionalProperties": false,
        "properties": {
          "data": {
            "nullable": true
          },
          "error": {
            "nullable": true,
            "type": "string"
          },
          "success": {
            "type": "boolean"
          },
          "usage": {
            "$ref": "#/components/schemas/Usage"
          }
        },
        "required": [
          "success",
          "data",
          "error"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "自动化端点使用的 API Key。"
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gptmail_session",
        "description": "Web Console 会话专用管理端点使用的会话 Cookie。"
      }
    }
  },
...
}